Make the input to `destroy` consistent with `create`
Now that domains in different regions can share a name, the `destroy`
command needs to know which region it is looking in.

Thus the input into `destroy` needs to be consistent with `create`
