+++ lib/cloudware/commands/domain/destroy.rb
@@ -4,11 +4,11 @@ module Cloudware
   module Commands
     module Domain
       class Destroy < Command
+        include Concerns::DomainInput
+
         def run
           d = Cloudware::Domain.new
-
-          options.name = ask('Domain name: ') if options.name.nil?
-          d.name = options.name.to_s
+          d.name = name
 
           run_whirly('Checking domain exists') do
             raise("Domain name #{options.name} does not exist") unless d.exists?
